Mark methods still needing work to eliminate dependency on dnlib Other minor refactorings
